apps/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dipeo-cli"
version = "2.0.0"
authors = [{name = "DiPeO Team"}]
description = "Command-line interface for DiPeO diagram operations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "gql[all]>=3.4.0",
    "aiohttp>=3.8.0",
    "websockets>=11.0",
    "dipeo-domain",
]
keywords = ["dipeo", "cli", "diagram", "workflow", "automation"]

[project.urls]
Homepage = "https://github.com/dipeo/dipeo-cli"
Repository = "https://github.com/dipeo/dipeo-cli"
Documentation = "https://github.com/dipeo/dipeo-cli/blob/main/README.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "ruff>=0.8.0",
]

[project.scripts]
dipeo = "dipeo_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
known_first_party = ["dipeo_cli"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "gql.*",
    "websockets.*",
]
ignore_missing_imports = true

[tool.ruff]
# Same as Black.
line-length = 88
target-version = "py38"

# Exclude generated files
exclude = [
    "__generated__",
    "generated",
    "migrations",
    "venv",
    ".git",
    "__pycache__",
    ".mypy_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
# Enable rules
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PLC", # pylint convention
    "PLE", # pylint error
    "PLW", # pylint warning
    "RUF", # ruff-specific rules
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "N818",  # exception names should end with Error (we have custom exceptions)
    "SIM108", # Use ternary operator (sometimes less readable)
]

[tool.ruff.lint.isort]
known-first-party = ["dipeo_cli", "src"]

[tool.ruff.lint.per-file-ignores]
# Tests can use assertions and relative imports
"tests/*" = ["S101", "TID252"]